# docs/swagger.yaml
definitions:
  api.Response:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  models.Domain:
    properties:
      created_at:
        type: string
      name:
        type: string
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
      uuid:
        type: string
    type: object
  models.Post:
    properties:
      content:
        type: string
      created_at:
        type: string
      domain:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
      uuid:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      domains:
        items:
          $ref: '#/definitions/models.Domain'
        type: array
      email:
        type: string
      name:
        type: string
      posts:
        items:
          $ref: '#/definitions/models.Post'
        type: array
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  requests.CreatePostRequest:
    properties:
      content:
        example: Echo is nice!
        type: string
      title:
        example: Echo
        type: string
    required:
    - content
    - title
    type: object
...
  requests.RegisterRequest:
    properties:
      email:
        example: username@gmail.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: P@ssw0r6
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  requests.UpdatePostRequest:
    properties:
      content:
        example: Echo is nice!
        type: string
      title:
        example: Echo
        type: string
    required:
    - content
    - title
    type: object
...
  responses.PostResponse:
    properties:
      content:
        example: Echo is nice!
        type: string
      title:
        example: Echo
        type: string
      username:
        example: John Doe
        type: string
      uuid:
        example: uuid
        type: string
    type: object
info:
  contact: {}
paths:
  /api/domain:
    get:
      consumes:
      - application/json
      description: Returns a sample response for listing domains.
      operationId: domain-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: List domains
      tags:
      - Domain Management
    post:
      consumes:
      - application/json
      description: Not implemented.
      operationId: domain-create
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Create domain
      tags:
      - Domain Management
...
  /api/post:
    get:
      consumes:
      - application/json
      description: Returns a list of posts for the specified domain.
      operationId: post-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.PostResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: List posts
      tags:
      - Post Management
    post:
      consumes:
      - application/json
      description: Creates a new post in the specified domain.
      operationId: post-create
      parameters:
      - description: Post creation data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Create post
      tags:
      - Post Management
  /api/post/{uuid}:
    delete:
      consumes:
      - application/json
      description: Removes a post by UUID from the specified domain.
      operationId: post-delete
      parameters:
      - description: Post UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete post
      tags:
      - Post Management
    get:
      consumes:
      - application/json
      description: Returns the details of a post by UUID within the specified domain.
      operationId: post-read
      parameters:
      - description: Post UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get post
      tags:
      - Post Management
    put:
      consumes:
      - application/json
      description: Modifies the details of a post by UUID within the specified domain.
      operationId: post-update
      parameters:
      - description: Post UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Post update data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Update post
      tags:
      - Post Management
...
  /api/user:
    get:
      consumes:
      - application/json
      description: Returns a list of users for the specified domain.
      operationId: user-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: List users
      tags:
      - User Management
    post:
      consumes:
      - application/json
      description: Creates a new user in the specified domain.
      operationId: user-create
      parameters:
      - description: User registration data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Create user
      tags:
      - User Management
...
  /register:
    post:
      consumes:
      - application/json
      description: Register creates a new user and domain based on the provided registration
        request.
      operationId: user-register
      parameters:
      - description: User's email, user's password
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterRequest'
      produces:
      - application/json
      responses: {}
      summary: Register a new user
      tags:
      - Account Actions
securityDefinitions:
  ApiKeyAuth:
    description: 'Provide the accessToken as a Bearer token: ''Bearer {accessToken}'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"